add system detection to stop the pain modifying admin.py again and again

+++ css3two_blog/admin.py
@@ -10,6 +10,7 @@ from django import forms
 from django.core.files.base import ContentFile
 import os
 from django.conf import settings
+import platform
 
 
 class MyModelForm(forms.ModelForm):
@@ -47,11 +48,12 @@ class BlogPostModelAdmin(admin.ModelAdmin):
         if obj:
             if obj.body:    # body有内容的时候才会更新md_file
                 filename = obj.filename
-                '''
                 if filename != 'no md_file':
-                    obj.md_file.delete(save=False)   # 部署的时候存在,可以正常删除文件
-                    obj.html_file.delete(save=False)
-                '''
+                    if platform.system() == "Windows":  # On Windows file can't be removed so leave it
+                        pass
+                    else:
+                        obj.md_file.delete(save=False)   # 部署的时候存在,可以正常删除文件
+                        obj.html_file.delete(save=False)
                 # 没有md_file就根据title创建一个, 但不能创建html因为obj.save()的时候会创建
                 obj.md_file.save(filename+'.md', ContentFile(obj.body), save=False)
                 obj.md_file.close()
